Reorganisation to avoid repeating code for MapKBB
